+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 - added properties to the `O_LIFT` object:
   - `wait_time`, which defines how long to wait after activation before beginning to move
   - `travel_distance`, which defines how many clicks the lift will travel vertically
+  - `speed`, which defines how many world units the lift will travel per frame
 - changed `Game mode selection` to allow hiding the dialog even after having completed the game (Gameplay → General → Game mode selection)
 - improved handling of dead enemies used as switch triggers in custom levels by not altering their activation status and by adding detection for having been exploded (#5682)
 - fixed Lara having an empty left holster if the option to remember guns between levels is used and she finished with the Desert Eagle (#5697)